Fixes: 1. Synchronize data between candidate and edge nodes 2. Fix server reverse connection to candidate/node timeouts, and http client connection timeout 3. Fix download failures when there's only one block and the block is in car format 4. Modify user activation code retrieval 5. Add bandwidth requirements for resource backup
New features:
1. Candidate and edge nodes submit PoW (Proof of Work)
2. Save user and node submitted PoW
3. Add user operations for asset management: CreateAsset, ListAssets, DeleteAsset, ShareAssets, and upload
4. Add user storage management interfaces: AllocateStorage, GetStorageSize, CreateAPIKey, GetAPIKeys, DeleteAPIKey
5. Implement DNS server to resolve assets to corresponding L1 nodes
6. Use multiple goroutines to detect NAT
7. Add registering private keys for candidate and edge nodes
8. Implement node scoring
9. User resource uploads
10. PoW summary on a timer
11. Node cache statistics
12. Node uplink and downlink traffic statistics

+++ node/device/impl.go
@@ -72,8 +72,8 @@ func (device *Device) GetNodeInfo(ctx context.Context) (types.NodeInfo, error) {
 	info.ExternalIP = device.publicIP
 	info.SystemVersion = version.String()
 	info.InternalIP = device.internalIP
-	info.BandwidthDown = float64(device.bandwidthDown)
-	info.BandwidthUp = float64(device.bandwidthUp)
+	info.BandwidthDown = device.bandwidthDown
+	info.BandwidthUp = device.bandwidthUp
 
 	mac, err := getMacAddr(info.InternalIP)
 	if err != nil {
